+++ src/Exception/TooManyRedirectsException.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace GuzzleHttp\Exception;
 
-class TooManyRedirectsException extends RequestException
+class TooManyRedirectsException extends ResponseException
 {
 }