+++ tests/Exception/ResponseExceptionTest.php
@@ -4,10 +4,8 @@ declare(strict_types=1);
 
 namespace GuzzleHttp\Tests\Exception;
 
-use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Exception\ResponseException;
-use GuzzleHttp\Exception\ResponseTimeoutException;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use PHPUnit\Framework\TestCase;
@@ -15,21 +13,19 @@ use Psr\Http\Client\NetworkExceptionInterface;
 use Psr\Http\Client\RequestExceptionInterface;
 
 /**
- * @covers \GuzzleHttp\Exception\ResponseTimeoutException
+ * @covers \GuzzleHttp\Exception\ResponseException
  */
-class ResponseTimeoutExceptionTest extends TestCase
+class ResponseExceptionTest extends TestCase
 {
     public function testHasRequestAndResponse(): void
     {
         $req = new Request('GET', '/');
         $res = new Response(200);
         $prev = new \Exception();
-        $e = new ResponseTimeoutException('foo', $req, $res, $prev, ['foo' => 'bar']);
+        $e = new ResponseException('foo', $req, $res, $prev, ['foo' => 'bar']);
 
-        self::assertInstanceOf(ResponseException::class, $e);
         self::assertInstanceOf(RequestException::class, $e);
         self::assertInstanceOf(RequestExceptionInterface::class, $e);
-        self::assertNotInstanceOf(ConnectException::class, $e);
         self::assertNotInstanceOf(NetworkExceptionInterface::class, $e);
         self::assertSame($req, $e->getRequest());
         self::assertSame($res, $e->getResponse());