+++ tests/Exception/ServerExceptionTest.php
@@ -1,9 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace GuzzleHttp\Tests\Exception;
 
 use GuzzleHttp\Exception\BadResponseException;
 use GuzzleHttp\Exception\RequestException;
+use GuzzleHttp\Exception\ResponseException;
 use GuzzleHttp\Exception\ServerException;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
@@ -24,12 +27,12 @@ class ServerExceptionTest extends TestCase
         $e = new ServerException('foo', $req, $res, $prev, ['foo' => 'bar']);
 
         self::assertInstanceOf(BadResponseException::class, $e);
+        self::assertInstanceOf(ResponseException::class, $e);
         self::assertInstanceOf(RequestException::class, $e);
         self::assertInstanceOf(RequestExceptionInterface::class, $e);
         self::assertNotInstanceOf(NetworkExceptionInterface::class, $e);
         self::assertSame($req, $e->getRequest());
         self::assertSame($res, $e->getResponse());
-        self::assertTrue($e->hasResponse());
         self::assertSame(500, $e->getCode());
         self::assertSame('foo', $e->getMessage());
         self::assertSame('bar', $e->getHandlerContext()['foo']);