+++ tests/PoolTest.php
@@ -7,6 +7,7 @@ namespace GuzzleHttp\Tests;
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\Exception\RequestException;
+use GuzzleHttp\Exception\ResponseException;
 use GuzzleHttp\Handler\CurlMultiHandler;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\HandlerStack;
@@ -321,8 +322,7 @@ class PoolTest extends TestCase
 
         foreach ($rejected as $reason) {
             self::assertInstanceOf(RequestException::class, $reason);
-            self::assertFalse($reason->hasResponse());
-            self::assertNull($reason->getResponse());
+            self::assertNotInstanceOf(ResponseException::class, $reason);
             self::assertArrayNotHasKey('http_code', $reason->getHandlerContext());
             self::assertArrayNotHasKey('header_size', $reason->getHandlerContext());
             self::assertArrayNotHasKey('content_type', $reason->getHandlerContext());